+++ sdk/paloaltonetworks/azure-resourcemanager-paloaltonetworks-ngfw/src/main/java/com/azure/resourcemanager/paloaltonetworks/ngfw/models/FirewallResourceUpdateProperties.java
@@ -33,9 +33,19 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
     private BooleanEnum isPanoramaManaged;
 
     /*
+     * Strata Cloud Managed: Default is False. Default will be CloudSec managed
+     */
+    private BooleanEnum isStrataCloudManaged;
+
+    /*
      * Panorama Configuration
      */
     private PanoramaConfig panoramaConfig;
+
+    /*
+     * Strata Cloud Manager Configuration, only applicable if Strata Cloud Manager is selected.
+     */
+    private StrataCloudManagerConfig strataCloudManagerConfig;
 
     /*
      * Associated Rulestack
@@ -129,6 +139,26 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
     }
 
     /**
+     * Get the isStrataCloudManaged property: Strata Cloud Managed: Default is False. Default will be CloudSec managed.
+     * 
+     * @return the isStrataCloudManaged value.
+     */
+    public BooleanEnum isStrataCloudManaged() {
+        return this.isStrataCloudManaged;
+    }
+
+    /**
+     * Set the isStrataCloudManaged property: Strata Cloud Managed: Default is False. Default will be CloudSec managed.
+     * 
+     * @param isStrataCloudManaged the isStrataCloudManaged value to set.
+     * @return the FirewallResourceUpdateProperties object itself.
+     */
+    public FirewallResourceUpdateProperties withIsStrataCloudManaged(BooleanEnum isStrataCloudManaged) {
+        this.isStrataCloudManaged = isStrataCloudManaged;
+        return this;
+    }
+
+    /**
      * Get the panoramaConfig property: Panorama Configuration.
      * 
      * @return the panoramaConfig value.
@@ -145,6 +175,29 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
      */
     public FirewallResourceUpdateProperties withPanoramaConfig(PanoramaConfig panoramaConfig) {
         this.panoramaConfig = panoramaConfig;
+        return this;
+    }
+
+    /**
+     * Get the strataCloudManagerConfig property: Strata Cloud Manager Configuration, only applicable if Strata Cloud
+     * Manager is selected.
+     * 
+     * @return the strataCloudManagerConfig value.
+     */
+    public StrataCloudManagerConfig strataCloudManagerConfig() {
+        return this.strataCloudManagerConfig;
+    }
+
+    /**
+     * Set the strataCloudManagerConfig property: Strata Cloud Manager Configuration, only applicable if Strata Cloud
+     * Manager is selected.
+     * 
+     * @param strataCloudManagerConfig the strataCloudManagerConfig value to set.
+     * @return the FirewallResourceUpdateProperties object itself.
+     */
+    public FirewallResourceUpdateProperties
+        withStrataCloudManagerConfig(StrataCloudManagerConfig strataCloudManagerConfig) {
+        this.strataCloudManagerConfig = strataCloudManagerConfig;
         return this;
     }
 
@@ -260,6 +313,9 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
         if (panoramaConfig() != null) {
             panoramaConfig().validate();
         }
+        if (strataCloudManagerConfig() != null) {
+            strataCloudManagerConfig().validate();
+        }
         if (associatedRulestack() != null) {
             associatedRulestack().validate();
         }
@@ -287,7 +343,10 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
         jsonWriter.writeJsonField("networkProfile", this.networkProfile);
         jsonWriter.writeStringField("isPanoramaManaged",
             this.isPanoramaManaged == null ? null : this.isPanoramaManaged.toString());
+        jsonWriter.writeStringField("isStrataCloudManaged",
+            this.isStrataCloudManaged == null ? null : this.isStrataCloudManaged.toString());
         jsonWriter.writeJsonField("panoramaConfig", this.panoramaConfig);
+        jsonWriter.writeJsonField("strataCloudManagerConfig", this.strataCloudManagerConfig);
         jsonWriter.writeJsonField("associatedRulestack", this.associatedRulestack);
         jsonWriter.writeJsonField("dnsSettings", this.dnsSettings);
         jsonWriter.writeArrayField("frontEndSettings", this.frontEndSettings,
@@ -320,8 +379,14 @@ public final class FirewallResourceUpdateProperties implements JsonSerializable<
                 } else if ("isPanoramaManaged".equals(fieldName)) {
                     deserializedFirewallResourceUpdateProperties.isPanoramaManaged
                         = BooleanEnum.fromString(reader.getString());
+                } else if ("isStrataCloudManaged".equals(fieldName)) {
+                    deserializedFirewallResourceUpdateProperties.isStrataCloudManaged
+                        = BooleanEnum.fromString(reader.getString());
                 } else if ("panoramaConfig".equals(fieldName)) {
                     deserializedFirewallResourceUpdateProperties.panoramaConfig = PanoramaConfig.fromJson(reader);
+                } else if ("strataCloudManagerConfig".equals(fieldName)) {
+                    deserializedFirewallResourceUpdateProperties.strataCloudManagerConfig
+                        = StrataCloudManagerConfig.fromJson(reader);
                 } else if ("associatedRulestack".equals(fieldName)) {
                     deserializedFirewallResourceUpdateProperties.associatedRulestack
                         = RulestackDetails.fromJson(reader);